Add urls and focus styles to address page tabs (#237)
This PR improved the accessibility and usability of Who Owns What by assigning a specific url path to each of the four tabs we include on the address page. Now users can tab through links to navigate between these pages as well as share each individual URL allowing another user to have a direct link to that page.