Utilize `DATABASE_URL` env variable in docker compose (#290)
This PR configures our docker-compose.yml file to reference the DATABASE_URL env variable when constructing the virtual environment, which allows the developer to dynamically override the database reference if they want to. It also updates our env.sample to include the correct default value for this env variable if the user follows the basic Docker instructions on the ReadMe.