Add typescript definitions to all data-related API responses (#294)
This PR fixes #258 by adding type definitions to all the data objects we receive from API calls to the database. It adds a new file called APIDataTypes.ts which includes these new typings as well as some pre-existing type infrastructure corresponding to data we input into these API calls.

The typings were made based on the expected responses from these APIs (in reference to the data in nycdb they retrieve). I made sure to leave a comment next to any type that doesn't match up with how the front-end will interpret it. I aim to deal with these cases when we address #297 .