Upgrade version if needed when user navigates to a new page. (#328)
This adds functionality that regularly checks to see if a new version is available, and if so, upgrades the user when they next click on a link, submit a form, or perform any other kind of page navigation.

The one annoying thing with this solution is that it actually reloads _after_ the page transition, which could be a bit jarring for the user, as they'll see the new page but then see it immediately reload.  Ideally we'd convince React Router to make all page transitions "hard" once we know an upgrade is available, but there doesn't seem to be a way to do this.