Added test for the bbl page params formatting function (#375)
This PR adds tests for our BBLPage component, which, for now, just consists of one test that checks whether we are appropriately reformatting the URL params that the component receives. This specific test makes sure that both url formats for our BBL direct links work— /bbl/4015640058 and /bbl/4/1564/58, for example.