Factor out createAddressPageRoutes() and add a 404 page. (#390)
This factors out a new function called `createAddressPageRoutes()` which cleans up the creation of links to sub-pages of the address page, making them more type-safe and DRY.

It also moves the `WhoOwnsWhatRoutes` component into `App.tsx` to avoid circular imports.  This is also more ideal: `routes.tsx` should be only concerned with routing pathnames, rather than what routes map to what components, making it as dependency-free as possible so that anything that imports it doesn't also end up importing every module in the codebase (this will come in handy with Jest's detection of dependent tests, as well as if/when we decide to do code splitting).

Unrelatedly, it also removes the double quotes around the word "owns" in the Twitter sharing message, because that was really weird.

Finally, it adds a very spartan 404 page, which fixes #389.  The text on the page is taken from the tenant platform so it's easy to localize.