Support property maps with no pins. (#406)
This fixes #405.  The problem arises when a portfolio only contains addresses with `null` latitude/longitude information, such as (at the time of this writing) "3070 Brighton 14th Street, Brooklyn".  This causes our bounding box calculation algorithm to return an _infinite_ bounding box which Mapbox really dislikes.

This resolves the situation by setting the default map bounds to be the NYC area instead of an infinite box.  I'm not sure if it's actually the most user-friendly fix since the user may be confused and wondering where the map pins are, but at least the entire page doesn't crash and the user can see other information about the properties?