Add new rentstab data to WOW sql queries (#412)
This PR updates our SQL code for creating the wow_bldgs table and our aggregate function for summary, in order to include the new 2019 rent stabilized unit counts we now have access to. In doing so, this PR also removes the percentchange field from our wow bldgs table set up, which was never used and would introduce unneeded complexity to recreate with the new data.