Add `allcontacts` field to registration_with_contacts table (#477)
In preparation for implementing Tahnee's new redesign of "Landlord Info" on the WOW Overview Tab, this PR adds a new column to the registration_with_contacts custom table that includes all HPD contact names and corporation names, with associated addresses, in one single json object.

In turn, this PR will subsequently add this new field called allcontacts to the end of the wow_bldgs table, since the SQL to generate the wow_bldgs table arbitrarily selects all columns from the registration_with_contacts table. Therefore, we should be able to start using this new column of data on the front-end right away.