Restructure how landlord contact info is displayed on the Overview Tab (#479)
This PR implements Tahnee's designs that more clearly organize and spell out all of the contact names and corporations associated with a building on WOW. While mostly a styling and html structure change, this PR makes use of #477 in that it properly shows all entity names with their associated business addresses (unlike before).

Beyond updating the Detail View on the overview tab, this PR changes the look and feel of the "useful links" section to make it more compact, which shows up on both the overview and timeline tabs.

One small win: the various expandable menus and text boxes in this PR were ALL implemented with the <summary><details> native HTML elements! No special JS needed! Also they are all accessible via tab focus.