Internationalize Landlord Contact Titles throughout Overview Tab (#484)
This PR internationalizes the titles for all HPD Contact types using a similar system to what we put in place with #461. This new functionality also allows us to automatically format these titles with proper spacing without having to do any string manipulation (I.e. HeadOfficer becomes Head Officer through this internationalization).

On the "Landlord Contact" section of the Overview tab, we also show the English version of the contact name alongside the translation, for clarity.