Restructure how datasets are selected on the Timeline Tab (#507)
This PR condenses the whole UI navigation of selecting a dataset to view on the Timeline Tab. Using Tahnee's designs, I implemented two accessible dropdown menus that allow the user to select a dataset and a "time span" from lists that can be arbitrarily long.

My implementation involved factoring out a new component called <Dropdown> which is shared between the header hamburger menu and this new feature. The new component has accompanying styles in Dropdown.scss as well.