Add DOB datasets to the timeline tab (#500)
This PR adds two new datasets to our indicatorhistory api call—DOB and ECB Violations—which subsequently adds this data as options on the Timeline Tab of Who Owns What. While this data comes from two sources, we encapsulate both datasets within one grouping called "DOB Violations" which includes both regular DOB Violations and ECB Violations. This PR is also a good illustration of all the moving parts that need to be adjusted or implemented in order to bring new datasets to the Timeline Tab in general, which highlights that the process of adding these is still quite complicated!