Generate a table of portfolios using networkx (#524)
This adds a new wow_portfolios table which contains disjoint portfolios that can be indexed by BBL. It's generated whenever the dbtool.py builddb command is run. Additionally, for development and debugging purposes, a new dbtool.py exportgraph command has been added, which outputs the rows of the table as a JSON file.

The new table's schema is defined in sql/create_portfolios_table.sql. Each row is a separate portfolio; one column contains a list of BBLs in the portfolio, another contains a list of landlord names in it, and the last contains the JSON graph structure of the portfolio itself.