Specify exact python version to use for CircleCI container (#528)
Soooo... upon trying to create a PR on who owns what this morning, we detected that our CircleCI build pipeline was failing, precisely when it was trying to download packages using Node 14 vs. the specified Node version 12 that we define within the Dockerfile as well as our engines list in package.json. The problem seemed to stem from the use of the CircleCi docker image called python:3.6-node, which according to the CircleCI docs pulls in Node 14 to accompany python 3.6.

Our working theory was that there was a python patch (Specifically 3.6.13 --> 3.6.14) that triggered CircleCI to update their image in a way that changed the associated Node version to 14. While we aren't 100% sure, specifying python version 3.6.13 in the reference to the CircleCI image in the config.yml file seemed to fix the issue entirely.