Calculate aggregation of portfolio data without external API call (#526)
This PR replaces the front-end use of our /aggregate API endpoint with a js function that calculates data for the Summary tab purely in the front-end and based on portfolio data already saved in client. This effectively removes the need for any external network request to render the Summary tab and all of it's aggregate calculations.

Notably, this change also involves updating our "state machine"— now that we are not making a network request on Summary tab page loading, there is no longer the concept of a network "error" state or network request "pending" state. Those states were removed.

Also, I went ahead and tracked how much time it took to run the new calculateAggDataFromAddressList() function. It was running in 1.5 milliseconds on my device, wahoo!