Update `wow_bldgs` in place without downtime (#551)
This PR fixes a bug where every morning when we update the wow_bldgs table on nycdb, we have about a 60-90 second period where there is no wow_bldgs table in the db. This happens because at the top of the create_bldgs_table.sql script, we DROP TABLE to the existing wow_bldgs table in the db, and then have to wait a minute or so for the new query to run that generates the updated table. This is particularly problematic, as dropping wow_bldgs here also drops the get_assoc_addrs_from_bbl function that is called every time a user searches an address on the live WOW site, according to this sample message:

NOTICE: drop cascades to function get_assoc_addrs_from_bbl(text,text)

You can see this sample rollbar error of this problem's affect on our users.

The fix? Creating a temporary wow_bldgs_temporary table, and then once it's been created, dropping the existing wow_bldgs table and renaming the temporary table back to wow_bldgs. This should reduce the amount of "downtime" to just a few milliseconds.