Add J51 & 421a exemption data to the `wow_bldgs` table (#550)
This PR adds two new columns to the wow_bldgs table that we use to source all the data presented on the "overview" tab. The two columns are:

yearstartedj51: the year that the landlord started receiving the J-51 tax exemption on the building (NULL if no exemption)
yearstarted421a: the year that the landlord started receiving the 421a tax exemption on the building (NULL if no exemption)
There are various other pieces of data related to these two exemption programs that we could include here, such as the number of years the exemption is set to last, and the amount of money the landlord is possibly saving from the exemption. However, I didn't want to add too much additional data to this table just yet—it's always easier to add more columns later than to remove superfluous columns.