Use `defaultLocale` global variable wherever possible (#564)
So, turns out that we have a global variable called defaultLocale that we aren't using everywhere. This PR makes sure we are referencing this variable (instead of the string "en") in all functions and components that need the default locale. For functions/components that are explicitly fetching the "English" translation of something, I left the string "en" as is.