clarify situation for addresses without portfolio (un/incomplete registration) (#615)
This PR changes the page shown when a building is not in a portfolio because the HPD registration is missing or incomplete. Previously it always said that a property is not registered, but actually in some cases they have a registration but it's missing important necessary to build a portfolio (ie. owner name/address). Now, the buildingInfo data collected for this page includes residential units and hpd registration dates, so we can check whether there is any HPD registration and will say "not registered" or "incomplete registration". This also changes the logic for what to display about the need to register - previously it was based on building class, but now it uses the number of residential units to more closely reflect HPD's guidance from their website (more details in PR comments #615)