Make sure users locale stays intact on address search (#626)
Oops! This PR fixes a weird bug where if a user searched and address from the homepage in Spanish, it would redirect them to an address page in English! Not sure how we missed this, but this PR ensures that the user's locale gets passed on correctly when they search an address.