Optimize the run time of the new WOWZA search query (#629)
I had a feeling that our new WOWZA search function was taking a while to run, often 2-4 seconds per search. While this isn't horrible, I did want to take a second to investigate any ways to cut down on run time.

I realized that the INNER JOIN WOW_BLDGS USING(BBL); line from the search function in address_portfolio.sql was the costliest line of code, so I rewrote the function without a costly join at all. There may be other room for improvement, but this fix cut down the run time to 1/4th of what it was before.