Reduce number of algolia network calls in LandlordSearch component (#630)
This PR addresses a very unfortunate problem with the Algolia instantsearch library: where any re-rendering of the component causes a duplicate network request to Algolia, which in turn costs us one of our precious search credits that we only get a certain amount of. So, this PR does some refactoring to reduce the number of times we re-render our custom LandlordSearch component to mitigate the problem.

Previously, our search bar was making 2 duplicate network requests on every user keystroke, which meant we were draining credits twice as fast as we should. I was able to optimize our rendering processes so that now, we make:

1 network request on initial renter of the search bar (this is expected)
2 network requests on typing in our first character
1 network request on each additional character entered in to the search bar
0 network requests on deleting a character from our search bar and showing cached search results
3 network requests on clearing the search bar

There are likely ways we can even further reduce how many network requests that we make, but this is a solution for now while we work out with Algolia support team WHY we the default behavior is to rerun a network request on each re-render of the component. Hopefully they can help us figure out how we can actually adjust the API calls themselves so we don't need to fiddle with our react implementation that much.