Add close button to big portfolio warning banner on map (#689)
This PR adds a close button for the big portfolio warning banner on the map and summary tab. For now it reappears on reload, but once JustFixNYC/justfix-ts#41 adds the storageFactory to our util package we can use sessionStorage to keep it closed for the duration of the user's session.

This also refactors the banner component to allow for other content, so we can use them for giving info about filter selections once we add that feature in addition to the portfolio size warning. The filter banners only shown in the images below as a preview for how it'll look later with multiple.