update amplitude sdk, track addr change & table sort (#696)
In anticipation of the release of the new portfolio filter feature, we want to start collecting a baseline of data on how often users are visiting other properties. This includes clicking on an address/bbl/detail from the portfolio table, and clicking on map points from the overview page. Also, this adds tracking for when a user sorts a column in the portfolio table (including event properties with the column name), so we can see if these interactions with the table also increase with the new filter features. This also changes us over to the new Amplitude SDK, matching what we have for org-site