use window size hook instead of one-time isMobile helper (#802)
The filter bar layout would break when changing from desktop to mobile on the same device because they filters get wrapped in a menu for mobile but the component wasn't re-rendering when it crossed that screen width threshold. This was because I was using the one-time Browser.isMobile, so in this PR I just switch that to our Helpers.useWindowSize hook so that it gets updated and the filters wrapper component is re-rendered as needed.