update portfolio table pagination bar styles (#806)
The contrast on the prev/next buttons in the portfolio table pagination bar were not meeting a11y standards, so while fixing that we made some other improvements to the design. The prev/next buttons now use icons, have clear hover and disabled states, and have 48px touch target for a11y. We also removed the ability to select a specific page (the page 1 of 2 is just text now).