show "loading" in place of streetview until screen dimensions known (#808)
In #793 we switched from dynamic to static streetview images for the address page. Because the images get easily distorted switching between mobile and desktop, we fetch a new version of the image based on window size, but on initial load there is a delay in getting the screen size and so it was defaulting to mobile size then quickly switching to desktop size, causing an extra API call.

This PR fixes the issue by just using a "loading" placeholder until the window size is known so we only make one request.