fix subscribed/verify/login state after adding building via login/signup #900
There is a weird bug with delayed user context update causing a flash, or at time persistently, showing the wrong state on the building page after adding a building via login/signup process. It would sometimes get stuck back on the beginning "check email" step when it should be subscribed/verify. And other attempted fixes would flash subscribed before verify or similar. In a rush to get this out for playbox UX testing we are just forcing a page reload at the end of login on building page and after closing the signup modal on building page. This ensures users see the correct state after the reload. Later we'll revisit and find a real solution