Nav bar style & text updates (#921)
* update address page url for detailAddr (#902)

replace the address page URL with the selected detailAdddr

Previously on WOW the address page url would always reflect the search address, and when you click on other properties in the portfolio from the map the detail page updates to that address but the URL remains the same. This caused confusion, since you could think you're copying the link for building you're currently viewing but its still the search address. Also it makes linking and using browser history navigation difficult or confusing. This will make other future changes easier, like getting back to your building after going through login/signup flow, or linking back to details from portfolio tba without needing to open a new tab or reloading the page.

* Simplified login/signup flow, removes modal on building page (#904)

This PR removes the entire on-building-page login/signup flow and modal. Now there is a single standalone login/signup page and when you're on a building page and click the "add building" button if you aren't logged in it just sends you to that page. We now make use of react-router's location state to set the building page's AddressRecord as state so we can have access to that from within the standalone login/signup flow. This allows us to still subscribe onSuccess of the login/register call, include the address subHeaders, provide a link back to the address page from the ending VerifyEmail step of those flows, or to automatically redirect users back to the building page on successful login for already verified users. We also make use of the same location state approach to set justSubscribed when landing back on the address page so that we can ensure it shows subscribed state immediately and we can know to show the page-level alert notifying the user that they are logged in and subscribed. Because the location state persists after reloads, we transfer the data to regular component state within a useEffect then clear the location state on both login and building page.

Also adds a loading state to all submit buttons in login/signup flow
Also ensures that emails are not broken by wrapping in login/signup flow

This also fixes the bug where after login/sign to subscribe to a building, the user context was out of sync and still showed the user as not logged in on the nav etc. It seems that this was due userContext.subscribe() overwriting the user values (email, verified) with the default undefined state because of the same kind of bug we had previous where an empty JustfixUser object is truthy, so you need to use !!user?.email in conditions instead (see shortcut story for details).

* Include origin to email verification request (#905)

* [sc-13805] send origin to as part of post data to verify_emails

* [sc-14357] after successful verification, show 'search for an address' link to home page

* [sc-14277] get repairs landing page is loc splash

* small updates based on feedback from early Playbook UX testing (#906)

A collection of small updates based on feedback from early playbook ux testing.

* update copy on verify success page [sc-14438]
* clarify password label for login vs register [sc-14439]
* clarify "verify" message on building page [sc-14446]
* add account link to login success message [sc-14447]
* remove newsletter from homepage [sc-14440]
* remove newsletter from all locations
* Update client/src/components/Login.tsx

* enable fullstory [sc-14379] (#908)

* Add analytics tags for email alerts (#907)

Adds all the Google Tag Manager (and Amplitude) events for our tracking in GA4.
Also makes updates to accommodate the user id and type now sent from auth provider, which we pass as params in our analytics events.
Corresponding auth-provider pr: JustFixNYC/auth-provider#39

* lingui

* nav bar compatibility with legacy mode

* exclude logo divider from svg styling for nav bar

* [sc-14476] remove logout button from navbar

---------

Co-authored-by: Maxwell Austensen <maxwell@justfix.org>
Co-authored-by: Maxwell Austensen <m.austensen@gmail.com>