add redirect login -> account when already logged in (#922)
This adds a redirect from the Login page to Account settings if you are already logged in. Also it shows a loading page while fetching the user object until we can decide whether to redirect or not. It uses a useEffect and extra state variable to track the redirect decision to avoid the problem of redirecting at the end of your login process before you can see the verification message.