allow bbl urls for specific parts of address page (#986)
On WOW we have a route for using BBL to get to the associated address page ( given /bbl/<bbl> it looks up the bbl to get the address and updates the url for the housenumber/streetname/borough version). But it has not been possible to link to specific tabs of the address page (eg. to the portfolio, timeline, summary) like you can with the address version of links

For GCE we want to link directly to the RS data on a building page's timeline tab as the source info for that data.

This PR adjusts the existing redirect setup for bbl -> address to allow for this.

Previously one component/page handled these redirects for the two variants for bbl routes - full bbl bbl/1234567890 and separated bbl/1/123/1234. This complicated things because now with additional parts to the path after /bbl/ if wouldn't know if that's a tab name like summary or a part of a split bbl. So I duplicated this page and have one for the separated bbl, and one for the full bbl.

We don't use the separated one ourselves, but I checked and we still get a lot of traffic via these style links, and I believe that's from old google results for specific addresses. (just fyi, I was able to use the sql console for our ga4 bigquery like for that recent data request to easily see events for page views with this link format)

I also had a problem of urls going to the wrong route because of the same issue of the patterns looking similar, so I've added some regex logic to the path when setting up the routes. react router uses this to make the paths, so I followed this convention. https://github.com/pillarjs/path-to-regexp/tree/v1.7.0

Since all the tests for this bbl page were related to the complications of the two different bbl patterns, I've just deleted the test file.

After all this was sorted out the main logic to allow the full links is just to join back on the extra parts of the path after the bbl -> address url conversion.