add API endpoint for dataset tracker info for sending alerts (#990)
We want to set up a dashboard display dataset update status and daily cron job script to check for out-of-date datasets, so this API endpoint returns all the information about datasets recorded via nycdb-k8s-loader and the manually created/maintained table of expected dataset schedules.

There may be some changes to this set up, but just wanted to get a first pass of this up and running now.