add dataset updat tracker table to /dev page (#991)
Uses the new API endpoint (#990) to add a simple table of all our datasets (nycdb and custom jobs) and when they were last updated and highlighting if any are late.

I kept the table extremely simple for now, without any interactively or extra data aside from what's in the API response.

Later I think it would be nice to add the ability to click on a row and reveal an extra row with a list of all the tables that dataset includes, what projects it's used in (wow, signature, etc.), and links to the the NYCDB wiki documentation page the sources datasets so we can check for problems on the open data portal. Unfortunately all this meta data isn't accessible from the NYCDB code itself, so I think I might just build this out manually in a static file within NYCDB and incorporate this later on.

(also fixes a bug in the api query for calculating if a dataset is late or not)