remove rows without bbl (#1008)
Somehow there were a ton of rows getting created that had no bbl, and so when aggregating all the related properties data by bbl (in the temp table immediately after this one being edited) there is one row for the null bbl that had an array being aggregated with millions of elements and it broke everything. Unfortunately I'm still not totally sure how these additional rows were getting created, but there are no duplicates among the other non-null bbl records. I could trace it down to this one join, and left a comment about it, but I don't see how this could happening, because when I do the same query without that last join there are no null-bbl rows, and the table being joined has no null-bbl rows, yet somehow after adding in that final join the result has a lot of null-bbl rows. Even more strange, is that this job was working just fine for a month or so then started failing just this weekend, though the null-bbl rows could have always been there but only recently did there become so many that it overflowed the maximum array length to raise an error.

But for now since the whole job is failing and GCE data isn't getting updated I wanted to get this fix in and can revisit later to try to better understand what's happening.