update email validation regex on signup to allow 2-char domains (#1024)
We got some feedback from a user that their me.com email was marked as invalid preventing them from signing up. We actually intended to allow 2-char domains but the way it was written (I think to allow subdomains?) it wasn't actually working that way.

I looked around at a bunch of example email validation regexes, and saw the advice that you can really only know for sure by attempting a send, but this one seemed to work for the cases we need and had a lot of up votes, so it seems good for our purposes.

https://stackoverflow.com/a/46181/7051239