update gce sql to fix c-of-o bbls with pluto match (#1030)
Following up on #1028 to further improve our updating of BBLs for CofO records. Now we check all records against PAD using BIN, not just those where the original BBL doesn't appear in PAD. And we also check against PLUTO to make sure we are matching to correct properties with residential units.

Examples:

https://whoownswhat.justfix.org/en/address/QUEENS/23-22/BROADWAY
bbl = 4005660136
If you go to this BBLs page on DOB (https://a810-bisweb.nyc.gov/bisweb/PropertyProfileOverviewServlet?boro=4&block=00566&lot=0136) and check the CofOs, there is one recorded in 2010 with BIN=4538508 (https://a810-bisweb.nyc.gov/bisweb/PropertyProfileOverviewServlet?requestid=2&bin=4538508).
We can find this CofO record in the DOB data (via FOIL for pre-2012) by looking for that BIN, but it is recorded with BBL=4005660037. This BBL 4005660037 doesn't not appear in pluto_latest. So it seems that the BBL for this BIN has changed since the CofO was first recorded. So we can join our CofO data to PAD (address databases that has BIN and BBL pairs) using BIN, we can get a current BBL associated with that bin, and we've confirmed that it gives BBL=4005660136. So now this address will show the correct "not covered" result on GoodCauseNYC because it finds a CofO after 2009.

https://whoownswhat.justfix.org/en/address/QUEENS/4545/CENTER%20BOULEVARD
bbl = 4000210020
Initially (in #1028) we were only doing this join with PAD and updating BBLs when the original BBL from the DOB records did not appear in PAD already and so definitely needed to be updated. However in this additional example we found that the CofO had a BBL that appears in PAD already, and even matched to pluto, but in pluto it's a record without a street address and have no residential units. So now we know what we need to check all CofO records against PAD to get current BBLs and also to check against PLUTO to make sure the BBL is present there as well.

In this case there are recent CofOs found on DOB's site for this BBL (https://a810-bisweb.nyc.gov/bisweb/PropertyProfileOverviewServlet?boro=4&block=00021&lot=0020) for BIN= 4541713 (https://a810-bisweb.nyc.gov/bisweb/COsByLocationServlet?requestid=0&allbin=4541713), but they weren't being found on GoodCauseNYC. The CofO record for BIN=4541713 has BBL=4000210001, and this BBL appears in PAD and pluto but with no units so doesn't seem correct. Joining with PAD using BIN= 4541713 we get BBL=4000210020 and this matches to pluto for a property with residential units. So now on GCE it will reflect that there has been a CofO since 2009 and correctly give a "not covered" result.

[sc-16412]